Son Nguyen - version 22.10.2023v1
https://itmsgroup-squad.monday.com/boards/4560723281/views/116055303/pulses/5354736424

The client advised that they don’t need to have any invoice for the treatment. We will track all items that are sold, but all invoice must be $0.00

The only invoices generated in the system are

1. Products sold on shop
2. $28.00 Prescriber fee

This means we need to change the main workflow.
After treatment is finished, no need to create invoice.
Just need to create aftercare and send email, then it’s complete